Refactor ConfigFile validation
ConfigFile now takes an optional list of required keys that it will use
to validate the file.  The file is valid if all keys are present.

This allows for this logic to be moved out of the client and into a
single, reusable place inside ConfigFile.

Since there's no need for the `has_key?` method to be exposed anymore, it
is also moved into the private methods as well.